split out fileresponse and xmlresponse into utils
